Add wait log
